libosmocc: Flag codec as not accepted instead of removing it
This is required if multiple SDP responses are returned. We don't want
to remove a codec with with the first SDP response, which might get
accepted by the second SDP response. Instead we flag the codec as
negotiated or not negotiated.

When early media is received, different codecs may be accepted as when
we receive the actual answer.