C-Netz SIM: Fix entering card numbers with leading zeroes
Numbers for the SIM card can be entered with zeroes in the front. This
will not change numbers to octal. The Bosch OF 7 phone will not accept
numbers in the phone book with less than 4 digits, so it is essential to
allow entering numbers with zeroes in front.

Also state that in the documents.